Refactor to separate codebase by application
