+++ requirements.txt
@@ -11,6 +11,7 @@ peakutils
 pymongo
 pyqt5>=5.9
 requests
+scikit-image
 shadow3>=23.1.4
 srwpy>=4.0.0b1
 tfs-pandas
