+++ sirepo_bluesky/__init__.py
@@ -1,26 +1,5 @@
-from ophyd import Signal
-
 from ._version import get_versions
 from .utils import prepare_re_env  # noqa: F401
 
 __version__ = get_versions()["version"]
 del get_versions
-
-
-class ExternalFileReference(Signal):
-    """
-    A pure software Signal that describe()s an image in an external file.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def describe(self):
-        resource_document_data = super().describe()
-        resource_document_data[self.name].update(
-            dict(
-                external="FILESTORE:",
-                dtype="array",
-            )
-        )
-        return resource_document_data
