+++ sirepo_bluesky/tests/test_srw_flyer.py
@@ -6,14 +6,14 @@ import pytest
 import vcr
 
 import sirepo_bluesky.tests
-from sirepo_bluesky.sirepo_bluesky import SirepoBluesky
-from sirepo_bluesky.sirepo_flyer import SirepoFlyer
+from sirepo_bluesky.common.sirepo_client import SirepoClient
+from sirepo_bluesky.srw.srw_flyer import SirepoFlyer
 
 cassette_location = os.path.join(os.path.dirname(sirepo_bluesky.tests.__file__), "vcr_cassettes")
 
 
 def _test_smoke_sirepo(sim_id, server_name):
-    sb = SirepoBluesky(server_name)
+    sb = SirepoClient(server_name)
     data, schema = sb.auth("srw", sim_id)
     assert "beamline" in data["models"]
